Refactor HTTP modules
