+++ lib/crawler/api/config.rb
@@ -128,6 +128,7 @@ module Crawler
         :compression_enabled,         # Enable/disable HTTP content compression
         :sitemap_discovery_disabled,  # Enable/disable crawling of sitemaps defined in robots.txt
         :head_requests_enabled,       # Fetching HEAD requests before GET requests enabled
+        :dynamic_content_enabled,     # Crawl dynamic content using HtmlUnit (rendered client-side by JS)
 
         # Sink lock retry settings
         :sink_lock_retry_interval,  # Interval in seconds to retry acquiring a sink lock
@@ -204,6 +205,7 @@ module Crawler
         extraction_rules: {},
         crawl_rules: {},
         purge_crawl_enabled: true,
+        dynamic_content_enabled: false,
         full_html_extraction_enabled: false,
 
         # Sink lock retry settings
